Loading Reference Cache left flags unset
loading refs from cache left a number of flags unset and structures uninitialized, calling Application.ProcessMessages from inside the closing of the in-place editor caused severe problems

+++ wbInterface.pas
@@ -36,7 +36,7 @@ type
 
 threadvar
   _wbProgressCallback : TwbProgressCallback;
-  wbCurrentTick       : Integer;
+  wbCurrentTick       : UInt64;
   wbCurrentAction     : string;
   wbStartTime         : TDateTime;
   wbShowStartTime     : Integer;
